add gentool for generating test sequences

+++ src/std/parse.h
@@ -1,6 +1,8 @@
 #ifndef FPLAYER_PARSE_H
 #define FPLAYER_PARSE_H
 
-void parseLong(const char *s, void *i, int n, long min, long max);
+#include <stddef.h>
+
+void parseLong(const char *s, void *i, size_t n, long min, long max);
 
 #endif//FPLAYER_PARSE_H
